211 update page agreement (mr:#257, is:#211)
