+++ .github/pull_request_template.md
@@ -12,24 +12,16 @@ This PR [adds/removes/fixes/replaces] the [feature/bug/etc].
 -   [ ] 🎨 Style
 -   [ ] 🧑‍💻 Code Refactor
 -   [ ] 🔥 Performance Improvements
--   [ ] ✅ Test
 -   [ ] 🤖 Build
 -   [ ] ⏩ Revert
--   [ ] 📈Data
+-   [ ] 📈 Data
 
 ## Related Tickets & Documents
 
-<!--
+<!-- Uncomment if relevant
 Please use this format link issue numbers: Fixes #123
 https://docs.github.com/en/free-pro-team@latest/github/managing-your-work-on-github/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword
 -->
-
-<!-- Uncomment if relevant
-## Added tests?
-
--   [ ] 👍 yes
--   [ ] 🙋 no, because I need help
--   [ ] 🙅 no, because they aren't needed -->
 
 <!-- Uncomment if relevant
 # Browser test?
@@ -51,6 +43,4 @@ https://docs.github.com/en/free-pro-team@latest/github/managing-your-work-on-git
 -   [ ] My code follows the style guidelines of this project
 -   [ ] I have performed a self-review of my code
 -   [ ] My changes generate no new warnings
--   [ ] I have added tests that prove my fix is effective or that my feature works
--   [ ] New and existing unit tests pass locally with my changes
 -   [ ] Any dependent changes have been merged and published in downstream modules
