--- src/app/components/agreements/agreements.component.html
@@ -1,108 +1,0 @@
-<div class='flex justify-center'>
-    <mat-card
-        class='h-min w-[320px] dark:[--mdc-elevated-card-container-color:#000000] tablette:w-[768px] desktop:w-[1700px]'
-    >
-        <mat-card-header>
-            <!--<icon
-                [color]='isDarkMode ? IconColorEnum.LIGHT : IconColorEnum.DARK'
-                [size]='40'
-                icon='agreement'
-                mat-card-avatar
-            ></icon>-->
-            <mat-card-title class='underline [--mat-card-title-text-size:25px]'>
-                {{ (wording | getAgreement).title | sentenceCase }}
-            </mat-card-title>
-        </mat-card-header>
-        <mat-card-content>
-            <div class='flex flex-col gap-8 pt-3'>
-                <div class='flex flex-col gap-2' id='agreement'>
-                    <div class='text-title font-bold'>
-                        {{ (wording | getAgreement).part1Title | sentenceCase }}
-                    </div>
-                    <ul class='flex list-disc flex-col gap-4'>
-                        <li
-                            *ngFor='
-                                let agreement of (wording | getAgreement)
-                                    .part1Text
-                            '
-                            class='ml-8 text-justify'
-                        >
-                            {{ agreement }}
-                        </li>
-                    </ul>
-                </div>
-
-                <div class='flex flex-col gap-2' id='tanks'>
-                    <div class='text-title font-bold'>
-                        {{ (wording | getAgreement).part2Title | sentenceCase }}
-                    </div>
-                    <div class='ml-8 break-all text-justify'>
-                        {{ (wording | getAgreement).part2Text }}
-                        <a
-                            class='inline-block text-link_light underline dark:text-link_dark'
-                            href='https://docs.google.com/spreadsheets/d/1VQhiiMaQqsH-zwiUaX-eH9NTBeBciRtqVCDck8WyTSU/edit?usp=sharing'
-                            rel='noreferrer'
-                            target='_blank'
-                        >
-                            https://docs.google.com/spreadsheets/d/1VQhiiMaQqsH-zwiUaX-eH9NTBeBciRtqVCDck8WyTSU/edit?usp=sharing
-                        </a>
-                    </div>
-                </div>
-
-                <div class='flex flex-col gap-2' id='permissions'>
-                    <div class='text-title font-bold'>
-                        {{ (wording | getAgreement).part3Title | sentenceCase }}
-                    </div>
-                    <div class='ml-8 break-all text-justify'>
-                        {{ (wording | getAgreement).part3Text }}
-                        <a
-                            class='inline-block text-link_light underline dark:text-link_dark'
-                            href='https://docs.google.com/spreadsheets/d/1f29HKQZ351By2dozf-m5daCGZN5y2gcO-ViVMKBeWl8/edit?usp=sharing'
-                            rel='noreferrer'
-                            target='_blank'
-                        >
-                            https://docs.google.com/spreadsheets/d/1f29HKQZ351By2dozf-m5daCGZN5y2gcO-ViVMKBeWl8/edit?usp=sharing
-                        </a>
-                    </div>
-                </div>
-
-                <div class='flex flex-col gap-2' id='officiers'>
-                    <div class='text-title font-bold'>
-                        {{ (wording | getAgreement).part4Title | sentenceCase }}
-                    </div>
-                    <ul class='flex list-disc flex-col gap-4'>
-                        <li
-                            *ngFor='
-                                let agreement of (wording | getAgreement)
-                                    .part4Text
-                            '
-                            class='ml-8 text-justify'
-                        >
-                            {{ agreement }}
-                        </li>
-                    </ul>
-                </div>
-
-                <div class='flex flex-col gap-2' id='organigramme'>
-                    <div class='text-title font-bold'>
-                        {{ (wording | getAgreement).part5Title | sentenceCase }}
-                    </div>
-                    <picture>
-                        <source
-                            srcset='assets/other/organigramme_clan.avif'
-                            type='image/avif'
-                        />
-                        <img
-                            alt='Organisation du clan LCHP'
-                            class='w-[320px] tablette:w-[750px]'
-                            height='794'
-                            ngSrc='assets/other/organigramme_clan.png'
-                            priority='0'
-                            width='1123'
-                        />
-                    </picture>
-                </div>
-            </div>
-        </mat-card-content>
-    </mat-card>
-</div>
